requirements.txt
--extra-index-url https://www.paddlepaddle.org.cn/packages/stable/cpu/
paddlepaddle==3.0.0
paddleocr
Pillow
pystray
pynput
pywebview
